requirements.txt
pyautogui>=0.9.54
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
PyQt5>=5.15.9
Pillow>=10.0.0
pywin32>=306
psutil>=5.9.0
pyinstaller>=6.0.0
